food_api_agent/requirements.txt
google-generativeai
python-dotenv
requests